finish dir structure for shadow dom example
