+++ examples/shadow-dom/counter-app/about-view/index.jade
