+++ examples/shadow-dom/counter-app/about-view/index.js
@@ -1,0 +1,12 @@
+// import from the same repo. in a different repo you'd use:
+// import { Component } from 'panel';
+import { Component } from '../../../../lib';
+
+import template from './index.jade';
+// import css from './index.styl';
+
+document.registerElement('about-view', class extends Component {
+  get config() {
+    return {template};
+  }
+});
