+++ examples/shadow-dom/counter-app/counter-view/index.jade
