+++ examples/shadow-dom/counter-app/index.jade
