+++ examples/shadow-dom/counter-app/index.js
@@ -2,11 +2,11 @@
 // import { Component } from 'panel';
 import { Component } from '../../../lib';
 
-import template from './app.jade';
-import aboutTemplate from './about.jade';
-import counterTemplate from './counter.jade';
+import './about-view';
+import './counter-view';
 
-import css from './app.styl';
+import template from './index.jade';
+import css from './index.styl';
 
 document.registerElement('counter-app', class extends Component {
   get config() {
@@ -26,28 +26,5 @@ document.registerElement('counter-app', class extends Component {
       css,
       useShadowDom: true,
     };
-  }
-});
-
-document.registerElement('about-view', class extends Component {
-  get config() {
-    return {template: aboutTemplate};
-  }
-});
-
-document.registerElement('counter-view', class extends Component {
-  get config() {
-    return {
-      helpers: {
-        decr: () => this.changeCounter(-1),
-        incr: () => this.changeCounter(1),
-      },
-
-      template: counterTemplate,
-    };
-  }
-
-  changeCounter(offset) {
-    this.update({count: this.state.count + offset});
   }
 });
